pyproject.toml
[tool.poetry]
name = "marsemseg"
version = "0.1.0"
description = "Marine segmentation"
authors = ["Sebastián García <segaracos@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.8"
mmsegmentation = "0.25.0"
mmcv-full = {url = "https://download.openmmlab.com/mmcv/dist/cu115/torch1.11.0/mmcv_full-1.5.3-cp38-cp38-manylinux1_x86_64.whl"}
torch = {url = "https://download.pytorch.org/whl/cu115/torch-1.11.0%2Bcu115-cp38-cp38-linux_x86_64.whl"}
torchvision = "0.12.0"
mmcls = "0.23.1"
openmim = "0.1.5"
requests = "^2.28.0"
addict = "2.4.0"
asttokens = "2.0.5"
backcall = "0.2.0"
click = "7.1.2"
colorama = "0.4.4"
cycler = "0.11.0"
debugpy = "1.6.0"
decorator = "5.1.1"
entrypoints = "0.4"
executing = "0.8.3"
fonttools = "4.33.3"
importlib-metadata = "4.11.4"
ipykernel = "6.13.1"
ipython = "8.4.0"
jedi = "0.18.1"
jupyter-client = "7.3.4"
jupyter-core = "4.10.0"
kiwisolver = "1.4.2"
Markdown = "3.3.7"
matplotlib = "3.5.2"
matplotlib-inline = "0.1.3"
model-index = "0.1.11"
nest-asyncio = "1.5.5"
opencv-python = "4.6.0.66"
ordered-set = "4.1.0"
packaging = "21.3"
pandas = "1.4.2"
parso = "0.8.3"
pexpect = "4.8.0"
pickleshare = "0.7.5"
prettytable = "3.3.0"
prompt-toolkit = "3.0.29"
psutil = "5.9.1"
ptyprocess = "0.7.0"
pure-eval = "0.2.2"
Pygments = "2.12.0"
pyparsing = "3.0.9"
python-dateutil = "2.8.2"
pytz = "2022.1"
PyYAML = "6.0"
pyzmq = "23.1.0"
stack-data = "0.2.0"
tabulate = "0.8.9"
tornado = "6.1"
traitlets = "5.2.2.post1"
wcwidth = "0.2.5"
yapf = "0.32.0"
zipp = "3.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

packages = [
    { include = ["src/*"] }
]

